+++ smbprotocol/open.py
@@ -467,6 +467,8 @@ class SMB2CreateResponse(Structure):
                 smbprotocol.create_contexts.SMB2CreateContextRequest()
             data = create_context.unpack(data)
             context_list.append(create_context)
+            # Manually make sure the final padding is present
+            create_context['padding2'] = b"\x00" * create_context._padding2_size(create_context)
             last_context = create_context['next'].get_value() == 0
 
         return context_list
@@ -932,7 +934,9 @@ class Open(object):
 
     def create(self, impersonation_level, desired_access, file_attributes,
                share_access, create_disposition, create_options,
-               create_contexts=None, send=True):
+               create_contexts=None,
+               oplock_level=RequestedOplockLevel.SMB2_OPLOCK_LEVEL_NONE,
+               send=True):
         """
         This will open the file based on the input parameters supplied. Any
         file open should also be called with Open.close() when it is finished.
@@ -966,6 +970,7 @@ class Open(object):
         opening files. More details on create context request values can be
         found here https://msdn.microsoft.com/en-us/library/cc246504.aspx.
 
+        :param oplock_level: The requested oplock level of the request.
         :param send: Whether to send the request in the same call or return the
             message to the caller and the unpack function
 
@@ -975,6 +980,7 @@ class Open(object):
             it is a Structure defined in create_contexts.py
         """
         create = SMB2CreateRequest()
+        create['requested_oplock_level'] = oplock_level
         create['impersonation_level'] = impersonation_level
         create['desired_access'] = desired_access
         create['file_attributes'] = file_attributes
@@ -1017,6 +1023,11 @@ class Open(object):
         response = self.connection.receive(request)
         create_response = SMB2CreateResponse()
         create_response.unpack(response['data'].get_value())
+
+        # Manually set the length so the debug log won't fail due to some servers returning a padded value which is not
+        # reflected in the padding2 of the context response.
+        create_response['create_contexts_length'] = len(create_response['buffer'])
+
         self._connected = True
         log.debug(str(create_response))